Move downloadFASTA into its own file
