+++ public/js/report.js
@@ -9,22 +9,7 @@ import HSP from './hsp';
 
 import SequenceModal from './sequence_modal';
 import showErrorModal from './error_modal';
-
-/**
- * Dynamically create form and submit.
- */
-var downloadFASTA = function (sequence_ids, database_ids) {
-    var form = $('<form/>').attr('method', 'post').attr('action', 'get_sequence');
-    addField('sequence_ids', sequence_ids);
-    addField('database_ids', database_ids);
-    form.appendTo('body').submit().remove();
-
-    function addField(name, val) {
-        form.append(
-            $('<input>').attr('type', 'hidden').attr('name', name).val(val)
-        );
-    }
-};
+import downloadFASTA from './download_fasta';
 
 /**
  * Base component of report page. This component is later rendered into page's
